Major refactor of StickyHeader; added ScreenGeometry
StickyHeaders now behave the way typical sticky headers do across the
web. There is an anchor piece that already exists on the page, and when
the user scroll to it it becomes sticky. Also improved the modularity
of the code.

Added ScreenGeometry, a library for doing calculations on elements with
respect to the screen and screen size.