Added StickyPushable, synchronized headers
Headers now synchronize themselves with the DOM by polling. Not the
most elegant way to do this, but I can’t see a more light-weight way to
do it. Moving the element itself has limitations too, as it is limited
by wherever it was created in the DOM originally, which may prevent it
from moving around. I may experiment with using an empty placeholder to
allow the element to be moved seamlessly, thereby bypassing the need
for synchronization altogether.

There could be a potential issue with z-indexes if I move away from
this model, since a nested element will never escape its parent’s
clipping.

StickyPushable are elements which stay inline, have no cloning
mechanism, and are pushed by the top of the screen/sticky stack.