Significant Changes: MutationObservers, Cleanup
Made significant improvements to StickyHeader. They no longer use the
sleep check and instead use the MutationObserver when it is supported.
As a fallback method, when the StickyHeader switches to active state it
will synchronize. Removed the synchronization and sleep cycles, so the
StickyHeaders are completely event driven. StickyHeaders will now clean
themselves up when they discover their origin element has been removed